+++ test/runtests.jl
@@ -152,7 +152,6 @@ end
 @safetestset "Integrated soil and snow" begin
     include("integrated/soil_snow.jl")
 end
-
 @safetestset "Full land" begin
     include("integrated/full_land.jl")
 end