+++ NEWS.md
@@ -3,6 +3,7 @@ ClimaLand.jl Release Notes
 
 main
 -----
+- ![][badge-✨feature] Add slab lake model to integrated models PR[#1672](https://github.com/CliMA/ClimaLand.jl/pull/1672)
 
 v1.6.5
 -----
@@ -27,7 +28,6 @@ v1.6.3
 
 v1.6.2
 -----
-
 - Update compat to SurfaceFluxes 1.0 PR[#1668](https://github.com/CliMA/ClimaLand.jl/pull/1668)
 - Clean up pkgversion checks and set ClimaDiagnostics in compat to 0.3 PR[#1676](https://github.com/CliMA/ClimaLand.jl/pull/1676)